// src/app/core/model/countries.interface.ts
export interface SummaryCountry {
    name:           string;
    capital?:       string;
    region:         Region;
    population:     number;
    flag:           string;
}

export interface CountryDetail {
    name:           string;
    nativeName:     string;
    population:     number;
    region:         Region;
    subregion:      string;
    capital?:       string;
    topLevelDomain: string[];
    currencies?:    Currency[];
    languages:      Language[];
    flag:           string;
    borders?:       string[]; 
}
// ...
export interface Currency {
    code:   string;
    name:   string;
    symbol: string;
}

export interface Flags {
    svg: string;
    png: string;
}

export interface Language {
    iso639_1?:   string;
    iso639_2:    string;
    name:        string;
    nativeName?: string;
}

export enum Region {
    Africa = "Africa",
    Americas = "Americas",
    Antarctic = "Antarctic",
    AntarcticOcean = "Antarctic Ocean",
    Asia = "Asia",
    Europe = "Europe",
    Oceania = "Oceania",
    Polar = "Polar",
}